pyproject.toml
[project]
name = "gotify-mcp"
version = "0.1.0"
description = "MCP server for Gotify push notifications"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.12",
    "fastmcp>=2.3.4",
    "httpx>=0.28.1",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.30.0",
]

[project.scripts]
gotify-mcp-server = "gotify_mcp.server:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
